create functions for initializing and freeing resources in main, move path freeing function to its own file
