+++ lem_in.h
@@ -64,6 +64,7 @@ ssize_t			lem_compare_nodes(t_graph_node *n1, t_graph_node *n2);
 ssize_t			lem_edge_remaining_capacity(t_graph_edge *edge);
 ssize_t			lem_find_node(t_array *dst, t_graph_node *node);
 void			lem_free_graph(t_graph *graph);
+void			lem_free_path_combinations(t_array *path_combinations);
 
 int				*lem_optimise_path_use(t_array **paths_to_use,
 					t_array *path_combinations, int ants);
