+++ src/lem_free_graph.c
@@ -1,7 +1,7 @@
 #include "lem_in.h"
 #include <stdlib.h>
 
-ssize_t free_edge(void *data, size_t i)
+static ssize_t free_edge(void *data, size_t i)
 {
 	t_graph_edge	*edge;
 
@@ -10,7 +10,7 @@ ssize_t free_edge(void *data, size_t i)
 	return ((ssize_t)i);
 }
 
-ssize_t free_node(void *data, size_t i)
+static ssize_t free_node(void *data, size_t i)
 {
 	t_graph_node	*node;
 	t_node_attr		*attr;
