+++ src/main.c
@@ -1,15 +1,51 @@
 #include "lem_in.h"
 #include <stdlib.h>
 
+static int	error(char *msg)
+{
+	print_fd(2, "%s\n", msg);
+	return (1);
+}
+
 ssize_t free_string(void *str, size_t i)
 {
 	free(str);
 	return ((ssize_t)i);
 }
 
-static int	error(char *msg)
+static void	free_resources(t_graph *graph, t_parray *input, t_parray *output)
 {
-	print_fd(2, "%s\n", msg);
+	lem_free_graph(graph);
+	if (input)
+	{
+		parr_iter(input, free_string);
+		parr_free(input);
+	}
+	if (output)
+	{
+		parr_iter(output, free_string);
+		parr_free(output);
+	}
+}
+
+static ssize_t	init_resources(t_graph *graph, t_parray *input, t_parray *output)
+{
+	*graph = lem_init_graph();
+	if (graph_null(graph))
+		return (-1);
+	*input = parr_new(1);
+	if (parr_null(input))
+	{
+		lem_free_graph(graph);
+		return (-1);
+	}
+	*output = parr_new(1);
+	if (parr_null(output))
+	{
+		lem_free_graph(graph);
+		parr_free(input);
+		return (-1);
+	}
 	return (1);
 }
 
@@ -19,39 +55,21 @@ int	main(void)
 	t_parray	input;
 	t_parray	output;
 
-	graph = lem_init_graph();
-	if (graph_null(&graph))
-		return (error("Error"));
-	input = parr_new(1);
-	if (input.data == NULL)
+	if (!init_resources(&graph, &input, &output))
 		return (error("Error"));
 	if (lem_parse_input(&graph, &input) != 1)
 	{
-		lem_free_graph(&graph);
-		parr_iter(&input, free_string);
-		parr_free(&input);
+		free_resources(&graph, &input, &output);
 		return (error("Error on parsing input"));
 	}
-	output = parr_new(1);
-	if (parr_null(&output))
-		return (error("Error"));
 	if (lem_process_graph(&output, &graph) != 1)
 	{
-		lem_free_graph(&graph);
-		parr_iter(&input, free_string);
-		parr_free(&input);
-		parr_iter(&output, free_string);
-		parr_free(&output);
+		free_resources(&graph, &input, &output);
 		return (error("Error on processing graph"));
 	}
-	lem_free_graph(&graph);
 	parr_iter(&input, lem_print_string);
 	print("\n");
 	parr_iter(&output, lem_print_string);
-	parr_iter(&input, free_string);
-	parr_free(&input);
-	parr_iter(&output, free_string);
-	parr_free(&output);
-//	system("leaks lem_in");
+	free_resources(&graph, &input, &output);
 	return (0);
 }
